apps/backend/pyproject.toml
[project]
name = "ai-stack-backend"
version = "0.1.0"
description = "FastAPI backend for AI Stack with RAG capabilities"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [
    { name = "AI Stack Team" }
]

dependencies = [
    # FastAPI & Server
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.18",
    
    # Database
    "sqlalchemy[asyncio]>=2.0.45",
    "asyncpg>=0.30.0",
    "alembic>=1.17.2",
    
    # Validation & Settings
    "pydantic>=2.12.5",
    "pydantic-settings>=2.7.0",
    "email-validator>=2.2.0",
    
    # Cache & Queue
    "redis>=5.2.1",
    "celery>=5.6.1",
    
    # Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    
    # LLM Providers
    "openai>=1.58.1",
    "anthropic>=0.42.0",
    "google-generativeai>=0.8.3",
    "ollama>=0.4.4",
    
    # Vector Databases
    "qdrant-client>=1.12.1",
    
    # Embeddings & NLP
    "tiktoken>=0.8.0",
    
    # Observability
    "langfuse>=2.57.6",
    "prometheus-client>=0.21.1",
    "structlog>=24.4.0",
    
    # Utilities
    "httpx>=0.28.1",
    "aiofiles>=24.1.0",
    "python-dotenv>=1.0.1",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.1",
    "ruff>=0.8.4",
    "mypy>=1.14.1",
    "black>=24.10.0",
    "isort>=5.13.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
